+++ source/Tubeshade.Server/Pages/LocalizationExtensions.cs
@@ -1,21 +1,29 @@
-﻿using Tubeshade.Data.Preferences;
+﻿using System.Resources;
+using Tubeshade.Data.Media;
+using Tubeshade.Data.Preferences;
 using Tubeshade.Server.Resources;
 
 namespace Tubeshade.Server.Pages;
 
 internal static class LocalizationExtensions
 {
+    private static readonly ResourceManager Manager = SharedResources.ResourceManager;
+
+    extension(ExternalAvailability availability)
+    {
+        internal string? LocalizedName => Manager
+            .GetString($"Video_{nameof(ExternalAvailability)}_{availability.Name}");
+    }
+
     extension(DownloadVideos download)
     {
-        internal string? LocalizedName => SharedResources
-            .ResourceManager
+        internal string? LocalizedName => Manager
             .GetString($"Preferences_{nameof(DownloadVideos)}_{download.Name}");
     }
 
     extension(DownloadMethod method)
     {
-        internal string? LocalizedName => SharedResources
-            .ResourceManager
+        internal string? LocalizedName => Manager
             .GetString($"Preferences_{nameof(DownloadMethod)}_{method.Name}");
     }
 }